# docs/swagger.yaml
basePath: /
info:
  contact: {}
  description: router.org clone implemented with Go + Gin by Elif ERDEN.
  title: HTTPBin Clone
  version: "1.0"
paths:
...
  /stream/{n}:
    get:
      description: Streams N JSON objects line by line
      parameters:
      - description: Number of lines
        in: path
        name: "n"
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stream JSON lines
      tags:
      - response
...
schemes:
- http
swagger: "2.0"